pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Liz Stanley <lizmstanley@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
prompt-toolkit = "3.0.17"
python-dateutil = ">=2.8.2"
python-dotenv = "1.0.1"
requests = "^2.32.3"
simplejson = "3.19.2"
toml = "^0.10.2"
typing-extensions = ">=4.6.1"
urllib3 = ">=2.2.2"
unidecode = "^1.3.8"
validators = "^0.28.1"
yarg = "0.1.9"


[tool.poetry.group.test.dependencies]
pytest = "8.2.0"
pytest-asyncio = "0.23.6"
pytest-mock = "3.14.0"
pyhamcrest = "2.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
